+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased] 
+---------------------
+##### Added
+- Support for "picker" appearance on range question.
+
 [1.33.0] - 2019-08-02
 ---------------------
 ##### Changed